Moved auth validation into module
