+++ index.js
@@ -13,69 +13,9 @@ const {
   AuthError,
   createLambda,
 } = require('@graffiticode/graffiticode-compiler-framework');
+const { validate } = require('./src/auth');
 const jsonDiff = require("json-diff");
-function postAuth(path, data, resume) {
-  let encodedData = JSON.stringify(data);
-  var options = {
-    host: "auth.artcompiler.com",
-    port: "443",
-    path: path,
-    method: "POST",
-    headers: {
-      'Content-Type': 'text/plain',
-      'Content-Length': Buffer.byteLength(encodedData),
-    },
-  };
-  var req = https.request(options);
-  req.on("response", (res) => {
-    var data = "";
-    res.on('data', function (chunk) {
-      data += chunk;
-    }).on('end', function () {
-      try {
-        resume(null, JSON.parse(data));
-      } catch (e) {
-        console.log("ERROR " + data);
-        console.log(e.stack);
-      }
-    }).on("error", function () {
-      console.log("error() status=" + res.statusCode + " data=" + data);
-    });
-  });
-  req.end(encodedData);
-  req.on('error', function(err) {
-    console.log("ERROR " + err);
-    resume(err);
-  });
-}
-function count(token, count) {
-  postAuth("/count", {
-    jwt: token,
-    lang: "L" + langID,
-    count: count,
-  }, () => {});
-}
-const validated = {};
-function validate(token, resume) {
-  if (token === undefined) {
-    resume(null, {
-      address: "guest",
-      access: "compile",
-    });
-  } else if (validated[token]) {
-    resume(null, validated[token]);
-    count(token, 1);
-  } else {
-    postAuth("/validate", {
-      jwt: token,
-      lang: "L" + langID,
-    }, (err, data) => {
-      validated[token] = data;
-      resume(err, data);
-      count(token, 1);
-    });
-  }
-}
+
 const recompileItem = (id, host, resume) => {
   let protocol, url;
   if (host === "localhost") {
@@ -153,12 +93,22 @@ const test = () => {
 };
 // SHARED STOP
 
+function isError(err) {
+  if (err instanceof Error) {
+    return true;
+  }
+  if (Array.isArray(err) && err.length > 0) {
+    return true;
+  }
+  return false;
+}
+
 const compilerDefinition = {
   language: langID,
   compile: (code, data, config) => {
     return new Promise((resolve, reject) => {
       compiler.compile(code, data, (err, val) => {
-        if (err) {
+        if (isError(err)) {
           reject(err);
         } else {
           resolve(val);
@@ -169,7 +119,7 @@ const compilerDefinition = {
   auth: (token) => {
     return new Promise((resolve, reject) => {
       validate(token, (err, data) => {
-        if (err) {
+        if (isError(err)) {
           reject(err);
         } else if (data.access.indexOf('compile') === -1) {
           reject(new AuthError('User does not have compile access'));
